move delete orders to POST files (as the method was changed from DELETE to POST)
